1016: Move metric logic to seperate modules.
